feat(): draw headers panels and message boxes, display combination guide to improve user experience, switch type of player score from int to double

+++ src/fr/uge/azathro/model/PlayerState.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-public record PlayerState(String name, int totalScore, int handCount, Map<Planet, Integer> planetDrawed) {
+public record PlayerState(String name, double totalScore, int handCount, Map<Planet, Integer> planetDrawed) {
     public static final int HANDCOUNT = 4;
 
     public PlayerState {
